Fix domain bugs: partial-fill loss + Decimal-for-money exceptions
P0 bugs identified during post-resumption codebase audit. Both fixes
are domain-layer and would silently corrupt data once Phase 2 puts a
real exchange in play.

1. Order.mark_closed silently destroyed partial fills.
   Old behavior: filled_amount=0.3 on a 1.0 order set status='closed'
   with the remainder vanished, even though the order is still live on
   the Kraken book.

   Replaced with record_fill(filled_amount: Decimal). Semantics:
   - Takes the *cumulative* filled amount, matching Kraken's vol_exec
     field on order responses.
   - status -> 'closed' only when filled_amount == amount (full fill).
   - status stays 'open' for partial fills (order remains on book).
   - Rejects monotonic violations (fill amounts can't decrease).
   - Rejects fills against non-open orders.

   Six new/updated tests cover full close, partial-keeps-open,
   accumulating partials, monotonic violation, overfill, and the
   pending-status guard.

2. ExposureLimitExceeded, DailySpendCapExceeded, InsufficientBalance
   typed their amount fields as `float`. Balance.lock had to call
   float(amount) and float(self.available) to fit the signature - the
   one place a precise value matters (balance violation reporting) was
   the one place precision was being lost.

   Changed signatures to Decimal, removed the float() casts in
   Balance.lock. The misleading test_exception_message_with_floats
   test was deleted; passing floats now violates the type contract and
   we don't want to advertise the behavior.

Tests: 81 passing (was 79; +6 new record_fill cases, -2 retired
mark_closed tests, -1 deleted float-interop test). mypy clean. black
clean.

+++ src/wobblebot/domain/exceptions.py
@@ -4,6 +4,8 @@ These exceptions represent business rule violations and should be raised
 by domain models when invariants are broken.
 """
 
+from decimal import Decimal
+
 
 class WobbleBotDomainError(Exception):
     """Base exception for all domain errors."""
@@ -14,7 +16,7 @@ class WobbleBotDomainError(Exception):
 class ExposureLimitExceeded(WobbleBotDomainError):
     """Raised when an operation would exceed configured exposure limits."""
 
-    def __init__(self, current: float, limit: float, message: str | None = None):
+    def __init__(self, current: Decimal, limit: Decimal, message: str | None = None):
         self.current = current
         self.limit = limit
         default_msg = f"Exposure limit exceeded: {current} > {limit}"
@@ -24,7 +26,7 @@ class ExposureLimitExceeded(WobbleBotDomainError):
 class DailySpendCapExceeded(WobbleBotDomainError):
     """Raised when daily spend cap is exceeded."""
 
-    def __init__(self, spent_today: float, cap: float, message: str | None = None):
+    def __init__(self, spent_today: Decimal, cap: Decimal, message: str | None = None):
         self.spent_today = spent_today
         self.cap = cap
         default_msg = f"Daily spend cap exceeded: {spent_today} > {cap}"
@@ -50,7 +52,7 @@ class InvalidGridConfiguration(WobbleBotDomainError):
 class InsufficientBalance(WobbleBotDomainError):
     """Raised when attempting an operation with insufficient funds."""
 
-    def __init__(self, required: float, available: float, asset: str):
+    def __init__(self, required: Decimal, available: Decimal, asset: str):
         self.required = required
         self.available = available
         self.asset = asset
